translate the rest of the tests
