+++ lucet-runtime/lucet-runtime-tests/src/lib.rs
@@ -8,4 +8,3 @@ pub mod memory;
 pub mod stack;
 pub mod start;
 pub mod strcmp;
-pub mod timeout;
