--- lucet-runtime/tests/timeout.rs
@@ -1,3 +1,0 @@
-use lucet_runtime_tests::timeout_tests;
-
-timeout_tests!(lucet_runtime::MmapRegion);
